% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOOR_prob_v2.0.R
\name{calculate_DOORprob}
\alias{calculate_DOORprob}
\title{Calculate the DOOR probability}
\usage{
calculate_DOORprob(data, DOORvar, trtvar, DOORlevel, trt = "1", ctrl = "0")
}
\arguments{
\item{data}{dataframe that contains required variables}

\item{DOORvar}{Variable name of DOOR}

\item{trtvar}{Variable name of treatment}

\item{DOORlevel}{number of DOOR level pre-specified}

\item{trt}{Character label for treatment in trtvar}

\item{ctrl}{character label for control in trtvar}
}
\value{
list of DOOR probability outputs
}
\description{
Calculate the DOOR probability
}
