% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOOR_prob_v2.0.R
\name{calculate_DOORprob_fromsummary}
\alias{calculate_DOORprob_fromsummary}
\title{Calculate the DOOR probability from summary table}
\usage{
calculate_DOORprob_fromsummary(
  DOORtable,
  DOORvar,
  trtvar,
  DOORlevel,
  trt = "1",
  ctrl = "0"
)
}
\arguments{
\item{DOORtable}{A dataframe contains the frequency crosstable of DOOR variable and Treatment variable}

\item{DOORvar}{Variable name of DOOR}

\item{trtvar}{Variable name of treatment}

\item{DOORlevel}{number of DOOR level pre-specified}

\item{trt}{Character label for treatment in trtvar}

\item{ctrl}{character label for control in trtvar}
}
\description{
Calculate the DOOR probability from summary table
}
