% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CART_DOOR.R
\name{cartdoor}
\alias{cartdoor}
\title{Subgroup identification function}
\usage{
cartdoor(
  data,
  DOORvar,
  trtvar,
  covar,
  DOORlevel,
  trt = "1",
  ctrl = "0",
  minsize = 30,
  incre = 5,
  midgroup = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{data}{the data frame}

\item{DOORvar}{specify the DOORvar name such as "DOOR"}

\item{trtvar}{specify the treatment variable name such as "ARMN"}

\item{covar}{specify the covariate name such as "covariate"}

\item{DOORlevel}{numeric value of the DOOR level}

\item{trt}{treatment indicator}

\item{ctrl}{control indicator}

\item{minsize}{minimum size of the split}

\item{incre}{increament of the split}
}
\value{
a vector of the split range
}
\description{
Subgroup identification function
}
