% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOOR_to.dummy.R
\name{DOOR_to.dummy}
\alias{DOOR_to.dummy}
\title{Converging DOOR categorical variable to a 1/0 matrix}
\usage{
DOOR_to.dummy(DOOR, nlevel)
}
\arguments{
\item{DOOR}{A vector contains DOOR outcome for all subjects}

\item{nlevel}{defined by user: how many DOOR levels is used in analysis}
}
\value{
A binary matrix which contains the DOOR category information
}
\description{
Converging DOOR categorical variable to a 1/0 matrix
}
