% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_CIs.R
\name{getCI_po}
\alias{getCI_po}
\title{Calculate the pointwise confidence interval for DOOR probabilities for new dataset}
\usage{
getCI_po(fit.po, newdata, DOORlevel)
}
\arguments{
\item{fit.po}{the fitted proportional odds model}

\item{newdata}{the new dataset contains arm indicator and covariate}

\item{DOORlevel}{Total DOOR level K}
}
\value{
a dataframe contains the covariate, DOOR probability, and confidence interval
}
\description{
Calculate the pointwise confidence interval for DOOR probabilities for new dataset
}
