% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametric_CI_v2.1.R
\name{Halperin_CI}
\alias{Halperin_CI}
\title{Calculate DOOR prob confidence intervals under null hypotheses}
\usage{
Halperin_CI(
  data,
  DOORvar,
  trtvar,
  DOORlevel,
  trt = "1",
  ctrl = "0",
  alpha = 0.05,
  method = c("null", "alternative")
)
}
\arguments{
\item{data}{A dataframe contains DOOR variable and Treatment variable}

\item{DOORvar}{Variable name of DOOR}

\item{trtvar}{Variable name of treatment}

\item{DOORlevel}{number of DOOR level pre-specified}

\item{trt}{Character label for treatment in trtvar}

\item{ctrl}{character label for control in trtvar}

\item{alpha}{significance level}

\item{method}{calculate variance under null hypothesis or alternative hypothesis}
}
\value{
list of DOOR probability and CIs
}
\description{
Calculate the DOOR probability from summary table
}
