% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOOR_sum_v0.R
\name{print.DOORsum}
\alias{print.DOORsum}
\title{Print method for S3 object DOORsum}
\usage{
\method{print}{DOORsum}(obj)
}
\arguments{
\item{obj}{S3 object DOORsum}
}
\description{
Print method for S3 object DOORsum
}
