% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOOR_prob_v2.0.R
\name{summarize_DOOR}
\alias{summarize_DOOR}
\title{Summarize DOOR variable by treatment groups}
\usage{
summarize_DOOR(data, DOORvar, trtvar, DOORlevel, trt = "1", ctrl = "0")
}
\arguments{
\item{data}{A dataframe contains DOOR variable and Treatment variable}

\item{DOORvar}{Variable name of DOOR}

\item{trtvar}{Variable name of treatment}

\item{DOORlevel}{number of DOOR level pre-specified}

\item{trt}{Char label for treatment indicator}

\item{ctrl}{char label for control indicator}
}
\description{
Summarize DOOR variable by treatment groups
}
