% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_diff_DOORprob.R
\name{test_diff_doorprob}
\alias{test_diff_doorprob}
\title{Test the difference in DOOR probabilities by subgroups}
\usage{
test_diff_doorprob(
  data,
  DOORlevel,
  trtvar,
  DOORvar,
  grpvar,
  grp_labels = c(1, 2),
  trt = "1",
  ctrl = "0"
)
}
\arguments{
\item{data}{the dataframe contains all variables}

\item{DOORlevel}{Total DOOR level K}

\item{trtvar}{the variable name for treatment indicator}

\item{DOORvar}{the variable name for DOOR}

\item{grpvar}{the variable name for subgroup indicator}

\item{grp_labels}{the subgroup labels}

\item{trt}{the treatment label}

\item{ctrl}{the control label}
}
\value{
p_value the p-value for the difference in DOOR probabilities
}
\description{
Test the difference in DOOR probabilities by subgroups
}
